+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dish tags generation on save ([#32](https://github.com/kooznitsa/sampo/pull/32)).
 - Dish tags generation commands ([#33](https://github.com/kooznitsa/sampo/pull/33)).
 - Admin methods to create restaurant nearest stations ([#34](https://github.com/kooznitsa/sampo/pull/34)).
+- Filter list of dishes by tags ([#36](https://github.com/kooznitsa/sampo/pull/36)).
 
 ### Fixed